requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
anthropic==0.42.0
pydantic==2.10.4
PyYAML==6.0.2
python-multipart==0.0.20
python-dotenv==1.0.1
